backend/requirements.txt
flask==2.3.2
rembg==2.0.66
pillow==10.0.0
flask-cors==4.0.0